refactor: dead-code cleanup, perf fixes, and structure refactor from audit
Dead code:
- Delete unused UnifiedModalView, StatusBadge, ZoomColors
- Remove unwired terminal search cluster (Station.startSearch etc.) and its
  orphaned ghosttySearch* notifications
- Remove never-called methods: stopPolling, cancelDownload,
  makeNewBranchDialog, eventsSince/clearEvents, setFocus, createEphemeral,
  *ForTesting helpers, and the never-posted .repoViewDidChangeWorktree pair

Performance:
- SplitContainerView: live-resize uses the applyFramesOnly fast path
  (zoom-aware) instead of full layoutTree per frame
- TabCoordinator: branch refresh polls background tabs every 30s instead of
  forking git per repo every 5s
- DiffSyntaxHighlighter: cache compiled regexes; exclusion filtering via
  sorted binary search instead of O(n^2) scan

Structure:
- Extract PersistedStringMap; WorktreeTaskStore/WorktreeSailorTypeStore are
  now thin wrappers, and persist failures are logged
- Split GhosttyNSView and ZmxSessionRecovery out of Station.swift

Docs: fix stale CLAUDE.md names (AgentHead/SurfaceRegistry/TerminalSurface ->
ShipLog/StationRegistry/Station)

via [Happy](https://happy.engineering)
