+++ Sources/UI/Dialog/NewBranchDialog.swift
@@ -1,74 +1,5 @@
 import AppKit
 
-// MARK: - Zoom Colors
-/// Zoom 风格的颜色系统，支持深色/浅色模式
-enum ZoomColors {
-    // 背景色
-    static let dialogBackground = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x1A1A1A) : NSColor(hex: 0xFFFFFF)
-    }
-    
-    static let inputBackground = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x2D2D2D) : NSColor(hex: 0xF5F5F5)
-    }
-    
-    // 文字颜色
-    static let titleText = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0xFFFFFF) : NSColor(hex: 0x1A1A1A)
-    }
-    
-    static let bodyText = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0xCCCCCC) : NSColor(hex: 0x333333)
-    }
-    
-    static let secondaryText = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x999999) : NSColor(hex: 0x666666)
-    }
-    
-    static let placeholderText = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x666666) : NSColor(hex: 0x999999)
-    }
-    
-    // Zoom 蓝色（强调色）
-    static let zoomBlue = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x0E72ED) : NSColor(hex: 0x0E72ED)
-    }
-    
-    static let zoomBlueHover = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x2B8CF7) : NSColor(hex: 0x2B8CF7)
-    }
-    
-    static let zoomBluePressed = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x0959C9) : NSColor(hex: 0x0959C9)
-    }
-    
-    // 边框颜色
-    static let borderDefault = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x3D3D3D) : NSColor(hex: 0xE0E0E0)
-    }
-    
-    static let borderFocus = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x0E72ED) : NSColor(hex: 0x0E72ED)
-    }
-    
-    // 按钮颜色
-    static let secondaryButtonBg = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x3D3D3D) : NSColor(hex: 0xF0F0F0)
-    }
-    
-    static let secondaryButtonHover = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0x4D4D4D) : NSColor(hex: 0xE0E0E0)
-    }
-    
-    // 错误颜色
-    static let errorText = NSColor(name: nil) { appearance in
-        appearance.isDark ? NSColor(hex: 0xFF6B6B) : NSColor(hex: 0xDC2626)
-    }
-    
-    // 阴影
-    static let shadowColor = NSColor.black.withAlphaComponent(0.4)
-}
-
 // MARK: - Typography
 enum ZoomTypography {
     static let title = NSFont.systemFont(ofSize: 18, weight: .semibold)
